Refactor episode list to render synopsis
This required disabling automatic parsing of Markdown to HTML as the
synopsis regex was designed to match markdown, not HTML. After that, a
new Twig extension was added to auto convert the markdown. I think this
is a simpler option and avoids unnecessary code, or mixing of concerns
etc (or the possibility of it).