Add new show and episode properties
These are required both for both entities and both for properly
representing a full iTunes RSS feed. I'm not saying that this is the
right way. Perhaps it should be refactored to keep that information
separate, and I'm tending to think that's the case. But for now, it's a
start.