Refactor the retrieval of episodes
Previously all data was being retrieved from one cache, which may or may
not have caused a problem with information iteration. I'm actually
thinking it's not an issue. I know I should have added a test, but will
shortly. So a separate cache key is now used for each type of episode
listing. Secondly, the retrieval of past episodes was broken. It was
limiting the information before it was fully prepared, so it was
unintentionally limiting from the wrong end.