Add episode 6 details
I've split the interview with Gary Hockin in to two parts, as there was
so much material over the hour that we talked for. I didn't want to cut
any of it out, so decided to share it all, in two parts, so I kept to
the normal running time.