Add Mockery support
The aim here is to make it as simple as possible to mock objects, when
necessary, during testing. Sure, there's mocking support in PHPUnit, but
I've never found it something that I found either straight-forward nor
intuitive. Mockery, with a little bit of learning, I found to be a much
better solution.