Add a further test to ensure the past episodes are sorted correctly
The main aim in doing this is that I want to refactor the
getPastEpisodes method to use PHP's SplPriorityQueue class to store the
retrieved past episodes, as there's less code to write. Previously, there
was insufficient code coverage to validate that a refactor hadn't
broken anything. Naturally, this couldn't stay in place long.