Add a new class to automatically store episodes in reverse date order
The intent here is to not have the sort logic be done after the fact, in
a separate class, but in the list itself. Given that, and thanks to
@elazar's talk at NomadPHP in Oct 2015, I learned about SplHeap and
SplMaxHeap which have a compare method which can store the logic ,
allowing for the sorting to be done as the objects are added.